Migrate away from BlazeBlogger, integrate with SMF
This commit is a complete overhaul of the site's backend. From now on,
we will rely on SMF to generate our news stories.

As such, the previously available blog stuff in the sidebar becomes
useless. It's been replaced with a list of most recently updated threads
from the forum (selected boards only) and a search bar for the forum at
large.

I probably broke something and we'll unbreak it momentarily.