Treat the header image as part of the content, not presentation
There is a long debate to be had about which approach is correct from a
sane semantic Web perspective. Let's not have that argument for the sake
of everyone's mental health. Using an <img> tag lets us appropriately
alt the image for screen readers/other robots without having to rely on
dodgy visibility:hidden <h1> tags.