Introduce h1 tag to header, try to display alt text correctly when needed
It's not immediately clear to me whether putting an image with an alt attribute inside an <h1> is entirely semantically correct, but it appears to be common practice. We may need to consider alternatives if this turns out to nor work as intended.

This commit also attempts to make the alt text for the logo to at least be human-readable. It won't look pretty if the logo doesn't loud, but it's a step forward. In future, it may be worth reproducing the text element of the logo as stylised text, rather than rely on an image.