Make Descriptions Great Again
This commit allows for meta descriptions (as well as OpenGraph descriptions) to be customised by setting the $descr variable in individual files that rely on header.php. In case of a missing $descr, the old default is used. Ideally, every significant page should have a distinct description.

For the time being, I am setting custom descriptions for copying.php and library.php - the default works well enough for the homepage.